+++ Libplanet.Explorer.Executable/Program.cs
@@ -229,7 +229,10 @@ If omitted (default) explorer only the local blockchain store.")]
                     // FIXME: The appProtocolVersion should be fixed properly.
                     var swarmOptions = new SwarmOptions
                     {
-                        MaxTimeout = TimeSpan.FromSeconds(10),
+                        TimeoutOptions = new TimeoutOptions
+                        {
+                            MaxTimeout = TimeSpan.FromSeconds(10),
+                        },
                     };
 
                     swarm = new Swarm<NullAction>(
@@ -366,8 +369,8 @@ If omitted (default) explorer only the local blockchain store.")]
                 Console.Error.WriteLine("Bootstrapping.");
                 await swarm.BootstrapAsync(
                     seeds,
-                    5000,
-                    5000,
+                    TimeSpan.FromMilliseconds(5000),
+                    TimeSpan.FromMilliseconds(5000),
                     cancellationToken: cancellationToken
                 );
             }