+++ CHANGES.md
@@ -8,12 +8,35 @@ To be released.
 
 ### Deprecated APIs
 
+ -  (Libplanet.Net) Unused property `SwarmOptions.PollInterval` removed.
+    [[#1962]]
+
 ### Backward-incompatible API changes
 
  -  `BlockCompletion<TPeer, TAction>.Complete()` no longer accepts
     neither parmeter `TimeSpan singleSessionTimeout` nor
     `int millisecondsSingleSessionTimeout` to regulate a single session length.
     [[#1961]]
+ -  (Libplanet.Net) General API changes made to `Swarm<T>.BootstrapAsync()`,
+    `Swarm<T>.PreloadAsync()`, and `Swarm<T>.StartAsync()`.  [[#1962]]
+     -  `Swarm<T>.BootstrapAsync(IEnumerable<Peer>, double, double,
+        int, CancellationToken)` replaced with
+        `Swarm<T>.BootstrapAsync(IEnumerable<Peer>, int, CancellationToken)`
+        which uses default values provided by `SwarmOptions` for
+        `pingSeedTimeout` and `findPeerTimeout`.
+     -  `Swarm<T>.StartAsync(int, int, int CancellationToken)`
+        replaced with `Swarm<T>.StartAsync(CancellationToken)`
+        which uses default values provided by `SwarmOptions` for
+        `millisecondsDialTimeout`, `millisecondsBroadcastBlockInterval`
+        and `millisecondsBroadcastTxInterval`.
+     -  `Swarm<T>.PreloadAsync(IProgress<PreloadState>,
+        bool, long, CancellationToken)` overload method added and
+        `Swarm<T>.PreloadAsync(TimeSpan?, IProgress<PreloadState>, bool
+        long, CancellationToken)` changed to `Swarm<T>.PreloadAsync(
+        TimeSpan, IProgress<PreloadState>, bool, long, CancellationToken)`
+ -  (Libplanet.Net) `TimeoutOptions` property added to `SwarmOptions` with
+    all timeout related options moved from `SwarmOptions` to `TimeoutOptions`.
+    [[#1957], [#1962]]
 
 ### Backward-incompatible network protocol changes
 
@@ -25,8 +48,11 @@ To be released.
  -  Added `StoreLoaderAttribute` class.  [[#1359], [#1953], [#1955]]
  -  Added `TrieStateStore.Secure` property.  [[#1955]]
  -  Added `NameValueCollectionExtensions` static class.  [[#1955]]
- -  Added `TimeSpanExtensions.Multiply()` static method.  [[#1966]]
  -  Type support for `Guid` added to `DataModel`.  [[#1959], [#1960]]
+ -  `TimeoutOptions` class added.  [[#1957], [#1962]]
+ -  `SwarmOptions.BlockBroadcastInterval` and `SwarmOptions.TxBroadcastInterval`
+    properties added.  [[#1962]]
+ -  `TimeSpanExtensions.Multiply()` method added.  [[#1966]]
 
 ### Behavioral changes
 
@@ -47,7 +73,6 @@ To be released.
 ### CLI tools
 
  -  The following store URI schemes are deprecated:  [[#1573], [#1955]]
-
      - `default`: Use `default+file` instead.
      - `rocksdb`: Use `rocksdb+file` instead.
 
@@ -55,9 +80,11 @@ To be released.
 [#1573]: https://github.com/planetarium/libplanet/issues/1573
 [#1953]: https://github.com/planetarium/libplanet/issues/1953
 [#1955]: https://github.com/planetarium/libplanet/pull/1955
+[#1957]: https://github.com/planetarium/libplanet/issues/1957
 [#1959]: https://github.com/planetarium/libplanet/issues/1959
 [#1960]: https://github.com/planetarium/libplanet/pull/1960
 [#1961]: https://github.com/planetarium/libplanet/pull/1961
+[#1962]: https://github.com/planetarium/libplanet/pull/1962
 [#1966]: https://github.com/planetarium/libplanet/pull/1966